# doc/openapi.yaml
...
openapi: 3.1.0
info:
  title: REST API best practices
  description: This API provides best practices for REST API design.
  version: 1.0.0-oas3-oas3.1
  contact:
    name: Suresh Raju Pilli
    url: https://www.pillisureshraju.in/
    email: sureshraju.pilli@gmail.com
servers:
  - url: http://localhost:8000
    description: Testing server
  - url: https://military-kesley-psrandapps-d7f588a1.koyeb.app
    description: Main (Production) Server
...
paths:
  /pets/anyof:
    post:
      summary: Add a pet
      description: Add a pet which can be either a cat or a dog.
      tags:
        - Pets
      operationId: addPet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Cat'
                - $ref: '#/components/schemas/Dog'
              discriminator:
                propertyName: petType
            examples:
              catExample:
                summary: A cat example
                value:
                  pet:
                    petType: cat
                    name: Whiskers
                    favoriteToy: Ball
              dogExample:
                summary: A dog example
                value:
                  pet:
                    petType: dog
                    name: Buddy
                    breed: Golden Retriever
      responses:
        '200':
          description: Pet added successfully.
        '400':
          description: Bad Request - Invalid input data.
        '422':
          description: Unprocessable Entity - Validation error.
        '500':
          description: Internal Server Error - Something went wrong on the server.
  /shapes/oneof:
    post:
      summary: Create a shape (oneOf example)
      description: Accepts either a rectangle or circle with strict validation
      operationId: createShape
      tags:
        - Shapes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneOfShape'
            examples:
              validRectangle:
                summary: Valid Rectangle
                value:
                  shape:
                    shape_type: rectangle
                    color: "#FF5733"
                    name: "MyRectangle"
                    width: 200
                    height: 100
                    aspect_ratio: 2.0
              invalidRectangle:
                summary: Invalid Rectangle
                value:
                  shape:
                    shape_type: rectangle
                    color: "invalid-color-code"
                    name: ""
                    width: -10
                    height: 1001
                    aspect_ratio: -1.0
              validCircleMinFields:
                summary: A valid circle with minimum fields
                value:
                  shape:
                    shape_type: circle
                    name: "MyCircle"
                    radius: 50
              invalidCircle:
                summary: An invalid circle example
                value:
                  shape:
                    shape_type: circle
                    name: "NoCircle"
                    radius: -10
      responses:
        '200':
          description: Shape created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  shape:
                    $ref: '#/components/schemas/OneOfShape'
                  validation_details:
                    type: object
        '400':
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
        '422':
          description: Unprocessable Entity - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneOfShapeErrorResponse'
  /products/allof:
    post:
      summary: Create a product (allOf example)
      description: Creates a product that must satisfy all schema requirements
      operationId: createProduct
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteProduct'
            examples:
              validProduct:
                summary: A valid product example
                value:
                  id: AB123456
                  name: Premium Widget
                  price: 29.99
                  stock_count: 100
                  warehouse_location: A-12-34
                  reorder_point: 25
                  weight_kg: 1.5
                  dimensions_cm:
                    - 30
                    - 20
                    - 15
                  fragile: true
              invalidProduct:
                summary: An invalid product example
                value:
                  id: "123456"
                  name: Widget
                  price: 29.99
                  stock_count: 100
                  warehouse_location: A-12-34
                  reorder_point: 25
                  weight_kg: 1.5
                  dimensions_cm:
                    - 30
                    - 20
                    - 15
                  fragile: true
              fieldMissing:
                summary: Field missing in the request
                value:
                  id: AB123456
                  name: Premium Widget
                  price: 29.99
                  stock_count: 100
                  reorder_point: 25
                  weight_kg: 1.5
                  dimensions_cm:
                    - 30
                    - 20
                    - 15
                  fragile: true
      responses:
        '200':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product:
                    $ref: '#/components/schemas/CompleteProduct'
                  calculated_metrics:
                    type: object
                    properties:
                      volume_m3:
                        type: number
                      density_kg_m3:
                        type: number
                      reorder_value:
                        type: number
  /offsetitems:
    get:
      summary: Get a paginated list of items
      description: Paginated response of the items
      operationId: createItems
      tags:
        - Items
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: Page number (starting from 1)
        - name: size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 50
          description: Number of items per page (max 50)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffsetPaginatedResponse'
  /cursoritems:
      get:
        summary: Get paginated items
        description: Retrieve a list of items using cursor-based pagination
        operationId: getItems
        tags:
          - Items
        parameters:
          - name: limit
            in: query
            description: Number of items to return per page
            required: false
            schema:
              type: integer
              minimum: 1
              maximum: 100
              default: 10
          - name: cursor
            in: query
            description: Cursor for pagination. Leave empty for first page
            required: false
            schema:
              type: string
          - name: direction
            in: query
            description: Direction of pagination
            required: false
            schema:
              type: string
              enum: [next, previous]
              default: next
        responses:
          '200':
            description: Successful retrieval of items
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/CursorPaginatedResponse'
          '400':
            description: Invalid cursor provided
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '404':
            description: Cursor position not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
...
  /laptops:
    get:
      summary: List laptops
      description: Retrieve a list of laptops with pagination and verbosity control
      operationId: listAllLaptops
      tags:
        - Laptops
      parameters:
        - name: verbosity
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ResponseVerbosity'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of laptops
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/LaptopBase'
                    - $ref: '#/components/schemas/LaptopRegular'
                    - $ref: '#/components/schemas/LaptopExtended'
components:
  schemas:
    Cat:
      type: object
      properties:
        petType:
          type: string
          enum:
            - cat
        name:
          type: string
        favoriteToy:
          type: string
      required:
        - petType
        - name
    Dog:
      type: object
      properties:
        petType:
          type: string
          enum:
            - dog
        name:
          type: string
        breed:
          type: string
      required:
        - petType
        - name
...
    OneOfShape:
      type: object
      properties:
        shape:
          oneOf:
            - $ref: '#/components/schemas/Rectangle'
            - $ref: '#/components/schemas/Circle'
      required:
        - shape
...
    CompleteProduct:
      allOf:
        - $ref: '#/components/schemas/BaseProduct'
        - $ref: '#/components/schemas/InventoryItem'
        - $ref: '#/components/schemas/ShippingDetails'
...
    OffsetPaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OffsetItem'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        total_pages:
          type: integer
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: Invalid cursor
      required:
        - detail
    ResponseVerbosity:
      type: string
      enum:
        - minimum
        - regular
        - extended
      default: regular
    LaptopBase:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the laptop
        brand:
          type: string
        model:
          type: string
        price:
          type: number
          format: float
      required:
        - id
        - brand
        - model
        - price
...
    OneOfShapeErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              input:
                type: object
              ctx:
                type: object
                additionalProperties: true
      example:
        detail: 
          - type: "missing"
            loc: ["body", "shape", "Rectangle", "width"]
            msg: "Field required"
            input: 
              shape_type: "circle"
              name: "NoCircle"
              radius: -10
          - type: "missing"
            loc: ["body", "shape", "Rectangle", "height"]
            msg: "Field required"
            input: 
              shape_type: "circle"
              name: "NoCircle"
              radius: -10
          - type: "greater_than"
            loc: ["body", "shape", "function-after[calculate_circle_properties(), Circle]", "radius"]
            msg: "Input should be greater than 0"
            input: 
              value: -10
            ctx: 
              gt: 0